Overlay OpenSSL for linux/unix
When we dynamically link against OpenSSL, the bundles are not portable,
as CentOS and friends use other SO_NAMEs than Debian and friends. Also,
soft-fallback for later features such as host name verification is hard.

When we statically link, we might lack behind the OS, the binaries are
bigger, and the legal situation is less clear.

So we now support not linking at all but rather lookup all necessary
functions/symbols at runtime.

This can be disabled with SQSSL_OPENSSL_LINKED which effectively results
in the dynamically-linked behavior. (This is preferable for platform
builds, eg, debs and rpms)